move app review managers to subpackages
